+++ src/net/server_net.cpp
@@ -166,7 +166,7 @@ void ServerNetwork::setContestTime ( ushort time )
         }
 }
 
-void ServerNetwork::setQuestion ( int qnum )
+void ServerNetwork::setQuestion ( ushort qnum )
 {
         if ( m_round >= 3 ) {
                 contestant_list::iterator i = m_contestants.begin();
